+++ examples/interop/pos-sidechain-example-two/src/app/modules/react/schemas.ts
@@ -1,45 +1,11 @@
-/**
- * Parameters of the reactCrossChain CCM
- */
-export interface CCReactMessageParams {
-	/**
-	 * A number indicating the type of the reaction.
-	 */
-	reactionType: number;
-	/**
-	 * ID of the Hello message being reacted to.
-	 */
-	helloMessageID: string;
-	/** Optional field for data / messages. */
-	data: string;
-}
-
-/**
- * Parameters of the react reactCrossChain command
- */
-export interface CCReactCommandParams extends CCReactMessageParams {
-	/**
-	 * The chain ID of the receiving chain.
-	 *
-	 * `maxLength` and `minLength` are equal to 4.
-	 */
-	receivingChainID: Buffer;
-	/**
-	 * The fee for sending the CCM across chains.
-	 */
-	messageFee: bigint;
-}
-
-/**
- * Schema for the parameters of the reactCrossChain CCM
- */
+// Schema for the parameters of the reactCrossChain CCM
 export const CCReactMessageParamsSchema = {
-	/** The unique identifier of the schema. */
+	// The unique identifier of the schema.
 	$id: '/lisk/react/ccmParams',
 	type: 'object',
-	/** The required parameters for the CCM. */
+	// The required parameters for the CCM.
 	required: ['reactionType', 'helloMessageID', 'data'],
-	/** A list describing the required parameters for the CCM. */
+	// A list describing the required parameters for the CCM.
 	properties: {
 		reactionType: {
 			dataType: 'uint32',
@@ -58,16 +24,14 @@ export const CCReactMessageParamsSchema = {
 	},
 };
 
-/**
- * Schema for the parameters of the react reactCrossChain command
- */
+// Schema for the parameters of the react reactCrossChain command
 export const CCReactCommandParamsSchema = {
-	/** The unique identifier of the schema. */
+	// The unique identifier of the schema.
 	$id: '/lisk/react/ccCommandParams',
 	type: 'object',
-	/** The required parameters for the command. */
+	// The required parameters for the command.
 	required: ['reactionType', 'helloMessageID', 'receivingChainID', 'data', 'messageFee'],
-	/** A list describing the available parameters for the command. */
+	// A list describing the available parameters for the command.
 	properties: {
 		reactionType: {
 			dataType: 'uint32',
