+++ examples/interop/pos-sidechain-example-two/src/app/modules/react/types.ts
@@ -7,6 +7,27 @@ import {
 } from 'lisk-sdk';
 
 export type TokenID = Buffer;
+// Parameters of the reactCrossChain CCM
+export interface CCReactMessageParams {
+	/**
+	 * A number indicating the type of the reaction.
+	 */
+	reactionType: number;
+	/**
+	 * ID of the Hello message being reacted to.
+	 */
+	helloMessageID: string;
+	/** Optional field for data / messages. */
+	data: string;
+}
+
+// Parameters of the react reactCrossChain command
+export interface CCReactCommandParams extends CCReactMessageParams {
+	// The chain ID of the receiving chain.
+	receivingChainID: Buffer;
+	// The fee for sending the CCM across chains.
+	messageFee: bigint;
+}
 
 export interface InteroperabilityMethod {
 	getOwnChainAccount(methodContext: ImmutableMethodContext): Promise<OwnChainAccount>;
